Change stack_top and stack_size to size_t (#3)
This essentially eliminates any risk of overflow on 64-bit
architectures (realloc() will always fail first). It also fixes the
missing stdint.h.

The (size_t)-1 cast eliminates a common warning.

+++ json.c
@@ -462,7 +462,7 @@ enum json_type json_next(json_stream *json)
         json->next = 0;
         return next;
     }
-    if (json->ntokens > 0 && json->stack_top == -1)
+    if (json->ntokens > 0 && json->stack_top == (size_t)-1)
         return JSON_DONE;
     int c = next(json);
     if (json->stack == NULL)
